server/cmd/app/requirements.txt
Flask==0.12
Werkzeug==0.11.15
Pillow==5.1.0
celery==4.2.1
elasticsearch==6.2.0
flask_mysqldb==0.2.0
aip==1.0.0.5
redis